docs(storybook): theme + tooltip docs
